feature: logout

+++ nestjs-server/src/models/users/user.controller.ts
@@ -5,6 +5,9 @@ import {
   Get,
   Param,
   Post,
+  Put,
+  UnauthorizedException,
+  UseGuards,
   UseInterceptors,
   UsePipes,
   ValidationPipe,
@@ -13,6 +16,10 @@ import {
 import { UserService } from './user.service';
 import { CreateUserRequest } from './dto/create-user.dto';
 import { ResponseInterceptor } from 'src/common/interceptors/response.interceptor';
+import JwtAuthGuard from 'src/auth/guards/jwt-auth.guard';
+import { CurrentUser } from 'src/common/decorators/currentUser.decorator';
+import { User } from './user.interface';
+import { UpdateUserRequest } from './dto/update-user.dto';
 
 @UseInterceptors(ResponseInterceptor)
 @Controller('users')
@@ -45,8 +52,30 @@ export class UserController {
       skipMissingProperties: true,
     }),
   )
+  @Put()
+  @UseGuards(JwtAuthGuard)
+  async update(
+    @CurrentUser() user: User,
+    @Body() request: UpdateUserRequest,
+  ): Promise<object> {
+    console.log('update');
+    console.log(user);
+    return this.userService.updateUserById(user.id, request);
+  }
+
+  @UsePipes(
+    new ValidationPipe({
+      whitelist: true,
+      skipMissingProperties: true,
+    }),
+  )
   @Delete(':id')
-  async delete(@Param('id') id: number): Promise<object> {
+  @UseGuards(JwtAuthGuard)
+  async delete(
+    @CurrentUser() user: User,
+    @Param('id') id: number,
+  ): Promise<object> {
+    if (user.id !== id) throw new UnauthorizedException('Invalid credentials');
     return this.userService.deleteById(id);
   }
 }
